backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.4
python-dotenv==1.0.0
openai==1.3.5
python-multipart==0.0.6
celery==5.5.2
redis==6.0.0
supabase==2.0.3
yt-dlp==2024.3.10
beautifulsoup4==4.12.2
playwright==1.42.0
requests==2.31.0
instaloader==4.10.1
moviepy==1.0.3 